moving tests into integration
